Feat optimistic locking (#38)
* Remove unused dependencies & files

* Add explicit version fields to domain, dtos & data mounting; lock exception handling

* Implement lock exception mechanism & refactoring

+++ src/main/java/distributed/monolith/learninghive/service/AuthorityServiceImpl.java
@@ -23,7 +23,7 @@ public class AuthorityServiceImpl implements AuthorityService {
 
 	@Override
 	public boolean isNotLoggedUser(@Nullable Long userId) {
-		return userId != null && securityService.getLoggedUserId() != userId;
+		return userId != null && securityService.getLoggedUserId().longValue() != userId.longValue();
 	}
 
 	@Override
